Aruco working
Detecting aruco 4x4 markers.

The problem was not using the correct dictionary.

Ed we could use the 6x6 dictionary, but need to import a different dictionary into python and I'm not sure how to do that. Currently this detects quite well the 4x4_50 markers, which should be sufficient I reckon.